toolkit/components/passwordmgr/test/mochitest/test_password_field_autocomplete.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test basic login autocomplete</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: multiple login autocomplete

<script>
var chromeScript = runChecksAfterCommonInit();

let origin = window.location.origin;
addLoginsInParent(
  // login0 has no username, so should be filtered out from the autocomplete list.
  [origin, "http://autocomplete:8888", null, "", "user0pass", "", "pword"],
  [origin, "http://autocomplete:8888", null, "tempuser1", "temppass1", "uname", "pword"],
  [origin, "http://autocomplete:8888", null, "testuser2", "testpass2", "uname", "pword"],
  [origin, "http://autocomplete:8888", null, "testuser3", "testpass3", "uname", "pword"],
  [origin, "http://autocomplete:8888", null, "zzzuser4", "zzzpass4", "uname", "pword"]);
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 tests multiple matching logins -->
  <form id="form1" action="http://autocomplete:8888/formtest.js" onsubmit="return false;">
    <h1>Sign in</h1>
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">
    <button type="submit">Submit</button>
  </form>

  <form id="form2" action="http://autocomplete:8888/formtest.js" onsubmit="return false;">
    <input  type="text"       name="uname">
    <input  type="password"   name="pword" readonly="true">
    <button type="submit">Submit</button>
  </form>

  <form id="form3" action="http://autocomplete:8888/formtest.js" onsubmit="return false;">
    <input  type="text"       name="uname">
    <input  type="password"   name="pword" disabled="true">
    <button type="submit">Submit</button>
  </form>

</div>

<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Login Manager: multiple login autocomplete. **/

var uname = $_(1, "uname");
var pword = $_(1, "pword");

// Restore the form to the default state.
async function reinitializeForm(index) {
  // Using innerHTML is for creating the autocomplete popup again, so the
  // preference value will be applied to the constructor of
  // LoginAutoCompleteResult.
  let form = document.getElementById("form" + index);
  let temp = form.innerHTML;
  form.innerHTML = "";
  // eslint-disable-next-line no-unsanitized/property
  form.innerHTML = temp;

  await new Promise(resolve => {
    let observer = SpecialPowers.wrapCallback(() => {
      SpecialPowers.removeObserver(observer, "passwordmgr-processed-form");
      resolve();
    });
    SpecialPowers.addObserver(observer, "passwordmgr-processed-form");
  });

  await SimpleTest.promiseFocus(window);

  uname = $_(index, "uname");
  pword = $_(index, "pword");
  uname.value = "";
  pword.value = "";
  pword.focus();
}

function generateDateString(date) {
  let dateAndTimeFormatter = new Services.intl.DateTimeFormat(undefined,
                                                              { dateStyle: "medium" });
  return dateAndTimeFormatter.format(date);
}

const DATE_NOW_STRING = generateDateString(new Date());

// Check for expected username/password in form.
function checkACFormPasswordField(expectedPassword) {
  var formID = uname.parentNode.id;
  is(pword.value, expectedPassword, "Checking " + formID + " password is: " + expectedPassword);
}

function spinEventLoop() {
  return Promise.resolve();
}

add_task(async function setup() {
  listenForUnexpectedPopupShown();
});

add_task(async function test_form1_initial_empty() {
  await SimpleTest.promiseFocus(window);

  // Make sure initial form is empty.
  checkACFormPasswordField("");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

add_task(async function test_form2_password_readonly() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["security.insecure_field_warning.contextual.enabled", true],
    ["signon.autofillForms.http", true],
  ]});
  await reinitializeForm(2);

  // Trigger autocomplete popup
  synthesizeKey("KEY_ArrowDown"); // open
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is closed for a readonly field.");
});

add_task(async function test_form3_password_disabled() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["security.insecure_field_warning.contextual.enabled", true],
    ["signon.autofillForms.http", true],
  ]});
  await reinitializeForm(3);

  // Trigger autocomplete popup
  synthesizeKey("KEY_ArrowDown"); // open
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is closed for a disabled field.");
});

add_task(async function test_form1_enabledInsecureFieldWarning_enabledInsecureAutoFillForm() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["security.insecure_field_warning.contextual.enabled", true],
    ["signon.autofillForms.http", true],
  ]});
  await reinitializeForm(1);
  // Trigger autocomplete popup
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  let results = await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  let expectedMenuItems = ["This connection is not secure. Logins entered here could be compromised. Learn More",
                           "No username (" + DATE_NOW_STRING + ")",
                           "tempuser1",
                           "testuser2",
                           "testuser3",
                           "zzzuser4"];
  checkAutoCompleteResults(results, expectedMenuItems, "mochi.test", "Check all menuitems are displayed correctly.");

  synthesizeKey("KEY_ArrowDown"); // select insecure warning
  checkACFormPasswordField(""); // value shouldn't update just by selecting
  synthesizeKey("KEY_Enter");
  await spinEventLoop(); // let focus happen
  checkACFormPasswordField("");
});

add_task(async function test_form1_disabledInsecureFieldWarning_enabledInsecureAutoFillForm() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["security.insecure_field_warning.contextual.enabled", false],
    ["signon.autofillForms.http", true],
  ]});
  await reinitializeForm(1);

  // Trigger autocomplete popup
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  let results = await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  let expectedMenuItems = ["No username (" + DATE_NOW_STRING + ")",
                           "tempuser1",
                           "testuser2",
                           "testuser3",
                           "zzzuser4"];
  checkAutoCompleteResults(results, expectedMenuItems, "mochi.test", "Check all menuitems are displayed correctly.");

  synthesizeKey("KEY_ArrowDown"); // select first item
  checkACFormPasswordField(""); // value shouldn't update just by selecting
  synthesizeKey("KEY_Enter");
  await spinEventLoop(); // let focus happen
  checkACFormPasswordField("user0pass");
});

add_task(async function test_form1_enabledInsecureFieldWarning_disabledInsecureAutoFillForm() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["security.insecure_field_warning.contextual.enabled", true],
    ["signon.autofillForms.http", false],
  ]});
  await reinitializeForm(1);

  // Trigger autocomplete popup
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  let results = await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  let expectedMenuItems = ["This connection is not secure. Logins entered here could be compromised. Learn More",
                           "No username (" + DATE_NOW_STRING + ")",
                           "tempuser1",
                           "testuser2",
                           "testuser3",
                           "zzzuser4"];
  checkAutoCompleteResults(results, expectedMenuItems, "mochi.test", "Check all menuitems are displayed correctly.");

  synthesizeKey("KEY_ArrowDown"); // select insecure warning
  checkACFormPasswordField(""); // value shouldn't update just by selecting
  synthesizeKey("KEY_Enter");
  await spinEventLoop(); // let focus happen
  checkACFormPasswordField("");
});

add_task(async function test_form1_disabledInsecureFieldWarning_disabledInsecureAutoFillForm() {
  await SpecialPowers.pushPrefEnv({"set": [
    ["security.insecure_field_warning.contextual.enabled", false],
    ["signon.autofillForms.http", false],
  ]});
  await reinitializeForm(1);

  // Trigger autocomplete popup
  synthesizeKey("KEY_ArrowDown"); // open
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is closed with no AutoFillForms.");
});

</script>
</pre>
</body>
</html>
